l18
Впровадив авторизацію в Rails за допомогою рішення Pundit,, а також створив сервісний об'єкт "гістьовий користувач". Успадкування контролерів . Готовий додаток із поділом прав доступу за ролями.

+++ app/controllers/comments_controller.rb
@@ -2,11 +2,15 @@
 
 class CommentsController < ApplicationController
   include QuestionsAnswers
+
+  before_action :require_authentication, except: %i[show index]
   before_action :set_commentable!
   before_action :set_question
+  after_action :verify_authorized
 
   def create
     @comment = @commentable.comments.build comment_params
+    authorize(@comment)
 
     if @comment.save
       flash[:success] = t '.success'
@@ -19,6 +23,8 @@ class CommentsController < ApplicationController
 
   def destroy
     comment = @commentable.comments.find params[:id]
+    authorize(comment)
+
     comment.destroy
     flash[:success] = t '.success'
     redirect_to question_path(@question)
